requirements_new.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.2
Pillow>=8.2.0
opencv-python>=4.5.2
pyminitouch>=0.3.3
pynput>=1.7.3
PyQt5>=5.15.4
requests>=2.25.1
loguru>=0.5.3
schedule>=1.1.0
pyyaml>=6.0
colorama>=0.4.4
tqdm>=4.62.0